refdataset/requirements.txt
# Required for dataloader
torch==2.1.1+cu118
torchvision==0.16.1+cu118
numpy==1.26.3
Pillow==10.2.0
pycocotools==2.0.7
scikit-image==0.22.0
matplotlib==3.9.2
